backend/catalog/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import CategoryViewSet, ProductViewSet, ReviewViewSet, site_content, yookassa_webhook

router = DefaultRouter()
router.register(r'categories', CategoryViewSet)
router.register(r'products', ProductViewSet)
router.register(r'reviews', ReviewViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('site-content/', site_content, name='site-content'),
    path('payments/yookassa/', yookassa_webhook, name='yookassa-webhook'),
]
