requirements.txt
Django==5.0.1
djangorestframework==3.14.0
aiogram==3.4.1
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
django-cors-headers==4.3.1
django-filter==23.5
asgiref==3.7.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
yookassa==2.4.0
